requirements.txt
opencv-python
pytesseract
pillow
pyqt6
reportlab
python-docx
argparse

